requirements.txt
python-telegram-bot[standard]
python-dotenv
